remove the "S" from all inference methods
