+++ sbi/inference/trainers/npe/npe_a.py
@@ -26,7 +26,7 @@ from sbi.utils.sbiutils import (
 from sbi.utils.torchutils import BoxUniform, assert_all_finite, atleast_2d
 
 
-class SNPE_A(PosteriorEstimator):
+class NPE_A(PosteriorEstimator):
     def __init__(
         self,
         prior: Optional[Distribution] = None,
@@ -37,18 +37,23 @@ class SNPE_A(PosteriorEstimator):
         summary_writer: Optional[TensorboardSummaryWriter] = None,
         show_progress_bars: bool = True,
     ):
-        r"""SNPE-A [1].
+        r"""NPE-A [1].
 
         [1] _Fast epsilon-free Inference of Simulation Models with Bayesian Conditional
             Density Estimation_, Papamakarios et al., NeurIPS 2016,
             https://arxiv.org/abs/1605.06376.
 
-        This class implements SNPE-A. SNPE-A trains across multiple rounds with a
+        Like all NPE methods, this method trains a deep neural density estimator to
+        directly approximate the posterior. Also like all other NPE methods, in the
+        first round, this density estimator is trained with a maximum-likelihood loss
+        in the first round.
+
+        This class implements NPE-A. NPE-A trains across multiple rounds with a
         maximum-likelihood-loss. This will make training converge to the proposal
         posterior instead of the true posterior. To correct for this, SNPE-A applies a
         post-hoc correction after training. This correction has to be performed
-        analytically. Thus, SNPE-A is limited to Gaussian distributions for all but the
-        last round. In the last round, SNPE-A can use a Mixture of Gaussians.
+        analytically. Thus, NPE-A is limited to Gaussian distributions for all but the
+        last round. In the last round, NPE-A can use a Mixture of Gaussians.
 
         Args:
             prior: A probability distribution that expresses prior knowledge about the
