+++ sbi/inference/trainers/nre/__init__.py
@@ -1,0 +1,9 @@
+from sbi.inference.trainers.nre.bnre import BNRE
+from sbi.inference.trainers.nre.nre_a import NRE_A
+from sbi.inference.trainers.nre.nre_b import NRE_B
+from sbi.inference.trainers.nre.nre_c import NRE_C
+
+# Aliases
+AALR = NRE_A
+SRE = NRE = NRE_B
+CNRE = NRE_C
