+++ sbi/inference/trainers/nre/nre_c.py
@@ -7,12 +7,12 @@ import torch
 from torch import Tensor, nn
 from torch.distributions import Distribution
 
-from sbi.inference.snre.snre_base import RatioEstimator
+from sbi.inference.trainers.nre.nre_base import RatioEstimator
 from sbi.sbi_types import TensorboardSummaryWriter
 from sbi.utils.sbiutils import del_entries
 
 
-class SNRE_C(RatioEstimator):
+class NRE_C(RatioEstimator):
     def __init__(
         self,
         prior: Optional[Distribution] = None,
@@ -23,18 +23,18 @@ class SNRE_C(RatioEstimator):
         show_progress_bars: bool = True,
     ):
         r"""NRE-C[1] is a generalization of the non-sequential (amortized) versions of
-        SNRE_A and SNRE_B. We call the algorithm SNRE_C within `sbi`.
+        NRE_A and NRE_B. We call the algorithm NRE_C within `sbi`.
 
         NRE-C:
-        (1) like SNRE_B, features a "multiclass" loss function where several marginally
+        (1) like NRE_B, features a "multiclass" loss function where several marginally
             drawn parameter-data pairs are contrasted against a jointly drawn pair.
-        (2) like AALR/NRE_A, i.e., the non-sequential version of SNRE_A, it encourages
+        (2) like AALR/NRE_A, i.e., the non-sequential version of NRE_A, it encourages
             the approximate ratio $p(\theta,x)/p(\theta)p(x)$, accessed through
             `.potential()` within `sbi`, to be exact at optimum. This addresses the
-            issue that SNRE_B estimates this ratio only up to an arbitrary function
+            issue that NRE_B estimates this ratio only up to an arbitrary function
             (normalizing constant) of the data $x$.
 
-        Just like for all ratio estimation algorithms, the sequential version of SNRE_C
+        Just like for all ratio estimation algorithms, the sequential version of NRE_C
         will be estimated only up to a function (normalizing constant) of the data $x$
         in rounds after the first.
