--- sbi/inference/trainers/snre/__init__.py
@@ -1,9 +1,0 @@
-from sbi.inference.snre.bnre import BNRE
-from sbi.inference.snre.snre_a import SNRE_A
-from sbi.inference.snre.snre_b import SNRE_B
-from sbi.inference.snre.snre_c import SNRE_C
-
-# Aliases
-AALR = SNRE_A
-SRE = SNRE = SNRE_B
-CNRE = NREC = SNRE_C
